+++ src/SharedAssemblyInfo.fs
@@ -4,10 +4,10 @@ open System.Reflection
 [<assembly: AssemblyCompanyAttribute("Yaaf.FSharp.Scripting")>]
 [<assembly: AssemblyProductAttribute("Yaaf.FSharp.Scripting")>]
 [<assembly: AssemblyCopyrightAttribute("Yaaf.FSharp.Scripting Copyright © Matthias Dittrich 2015")>]
-[<assembly: AssemblyVersionAttribute("1.0.4")>]
-[<assembly: AssemblyFileVersionAttribute("1.0.4")>]
-[<assembly: AssemblyInformationalVersionAttribute("1.0.4")>]
+[<assembly: AssemblyVersionAttribute("1.0.5")>]
+[<assembly: AssemblyFileVersionAttribute("1.0.5")>]
+[<assembly: AssemblyInformationalVersionAttribute("1.0.5")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.0.4"
+    let [<Literal>] Version = "1.0.5"